feat: add middleware roles test
